feat(render): add node drawing functionality
## Me

> Sharing:
> - script.ls

Insert code at the position {{{}}} of #buffer, such that it complete the function draw-node. The function will use the current font to put a node.label string rounding by a rounded-square. node.x will be the left border of the rounded-square, and node.y will be the middle of the rounded-square

## Claude

```diff
@label@ -102,6 +102,13 @@
 draw-node = (node, ctx) !->
-    {{{}}
+    ctx.font = label-font
+    ctx.begin-path!
+    ctx.round-rect node.x, node.y - node.h/2, node.w, node.h, 8
+    ctx.stroke!
+    ctx.fill-text node.label, node.x + label-padding, node.y + node.text-height/2 - node.h/4
+    for child in node.children
+      ctx.begin-path!
+      ctx.move-to node.x + node.w, node.y
+      ctx.line-to child.x, child.y
+      ctx.stroke!
+      draw-node child, ctx
```